extract date utils
