--- src/utils/duration.ts
@@ -1,9 +1,0 @@
-import { differenceInYears, formatDuration } from "date-fns";
-
-export const getExperienceDuration = () => {
-  const today = new Date();
-  const init = new Date(2018, 1, 1);
-  const years = differenceInYears(today, init);
-
-  return formatDuration({ years });
-};
